# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop Documentation",
        "description": "Petshop API",
        "contact": {
            "email": "mmabdelgawad@outlook.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://petshop.local/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Returns user token",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all products",
                "description": "products filter",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Product Category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product Price",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Product Brand",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Product Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Product Sort (ASC or DESC)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Product Limit",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Product Page",
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserLogin": {
                "properties": {
                    "email": {
                        "description": "email",
                        "format": "string",
                        "default": "petshop@buckhill.com"
                    },
                    "password": {
                        "description": "password",
                        "format": "string",
                        "default": "password"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "400": {
                "description": "Bearer Token is missing",
                "content": {
                    "application/json": {
                        "schema": {}
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}